test(backend): 拆分 direct injection 测试到独立文件
将 TestGetOriginalDocumentsFromKnowledgeBase 等三个测试类
从 test_retrieval_service.py 移动到 test_direct_injection.py，
原文件从 1450 行降至 790 行，符合单文件 1000 行限制
